requirements.txt
pyTelegramBotAPI==4.10.0
python-decouple==3.8
gtts==2.5.3
